# api/assets/swagger.json
{
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "schemes": ["http"],
  "swagger": "2.0",
  "info": {
    "description": "Jovvix",
    "version": "0.0.1-alpha"
  },
  "host": "127.0.0.1",
  "basePath": "/api",
  "paths": {
...
    "/v1/final_score/user": {
      "get": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["FinalScore"],
        "summary": "Get a finalScore details for user.",
        "operationId": "RequestFinalScoreForUser",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "UserPlayedQuiz",
            "name": "user_played_quiz",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseFinalScoreForUser"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
    "/v1/images": {
      "post": {
        "consumes": ["multipart/form-data"],
        "schemes": ["http", "https"],
        "tags": ["FileUpload"],
        "summary": "Insert image in s3 bucket.",
        "operationId": "RequestInsertImage",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "QuizId",
            "name": "quiz_id",
            "in": "query",
            "required": true
          },
          {
            "type": "file",
            "x-go-name": "File",
            "name": "image-attachment",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseInsertImage"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "401": {
            "$ref": "#/responses/GenericResFailConflict"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
...
    "/v1/quizzes/{quiz_id}": {
      "delete": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["Quiz"],
        "summary": "Delete quiz that created by user (if no active quiz is present).",
        "operationId": "DeleteQuizById",
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseOkWithMessage"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
...
    "/v1/quizzes/{quiz_id}/questions": {
      "get": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["Question"],
        "summary": "List all questions of quiz with `is_active_quiz_present` and `quiz_played_count`.",
        "operationId": "RequestListQuestionByQuizId",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "QuizId",
            "name": "quiz_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseListQuestionByQuizId"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
    "/v1/quizzes/{quiz_id}/questions/{question_id}": {
      "get": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["Question"],
        "summary": "Get question and thier options with answer.",
        "operationId": "RequestGetQuestionById",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "QuizId",
            "name": "quiz_id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "QuestionId",
            "name": "question_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseGetQuestionById"
          },
          "401": {
            "$ref": "#/responses/GenericResFailConflict"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      },
      "put": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["Question"],
        "summary": "Update question and thier options with answer.",
        "operationId": "RequestUpdateQuestionById",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "QuizId",
            "name": "quiz_id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "QuestionId",
            "name": "question_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseOkWithMessage"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "401": {
            "$ref": "#/responses/GenericResFailConflict"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      },
      "delete": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["Question"],
        "summary": "Delete question only if no active quiz is present.",
        "operationId": "RequestDeleteQuestionById",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "QuizId",
            "name": "quiz_id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "QuestionId",
            "name": "question_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseOkWithMessage"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "401": {
            "$ref": "#/responses/GenericResFailConflict"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
...
    "/v1/user_played_quizes": {
      "get": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["UserPlayedQuiz"],
        "summary": "List played quiz by user.",
        "operationId": "UserPlayedQuiz",
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseListUserPlayedQuizes"
          },
          "401": {
            "$ref": "#/responses/GenericResFailConflict"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "QuestionAnalytics": {
      "type": "object",
      "properties": {
        "correct_answer": {
          "type": "string",
          "x-go-name": "CorrectAnswer"
        },
        "duration_in_seconds": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "DurationInSeconds"
        },
        "options": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Options"
        },
        "options_media": {
          "type": "string",
          "x-go-name": "OptionsMedia"
        },
        "points": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Points"
        },
        "question": {
          "type": "string",
          "x-go-name": "Question"
        },
        "question_id": {
          "type": "string",
          "x-go-name": "QuestionId"
        },
        "question_media": {
          "type": "string",
          "x-go-name": "QuestionsMedia"
        },
        "question_type": {
          "type": "string",
          "x-go-name": "QuestionType"
        },
        "question_type_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "QuestionTypeID"
        },
        "raw_options": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "uint8"
          },
          "x-go-name": "RawOptions"
        },
        "resource": {
          "type": "string",
          "x-go-name": "Resource"
        }
      },
      "x-go-package": "github.com/Improwised/jovvix/api/pkg/structs"
    },
...
  },
  "responses": {
    "GenericResError": {
      "description": "Unexpected error occurred",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "x-go-name": "Data"
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "status": {
            "type": "string",
            "enum": ["error"],
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "GenericResFailConflict": {
      "description": "Server understand request but refuse to authorize it",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "enum": ["fail"],
            "x-go-name": "Status"
          }
        }
      }
    },
    "GenericResFailNotFound": {
      "description": "Fail due to resource not exists",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "enum": ["fail"],
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ResponseFinalScoreForUser": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "firstname": {
                  "type": "string",
                  "x-go-name": "FirstName"
                },
                "img_key": {
                  "type": "string",
                  "x-go-name": "ImageKey"
                },
                "rank": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "Rank"
                },
                "response_time": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "ResponseTime"
                },
                "score": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "Score"
                },
                "username": {
                  "type": "string",
                  "x-go-name": "UserName"
                }
              }
            },
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ResponseGetQuestionById": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "correct_answer": {
                "type": "string",
                "x-go-name": "CorrectAnswer"
              },
              "duration_in_seconds": {
                "type": "integer",
                "format": "int64",
                "x-go-name": "DurationInSeconds"
              },
              "options": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                },
                "x-go-name": "Options"
              },
              "options_media": {
                "type": "string",
                "x-go-name": "OptionsMedia"
              },
              "points": {
                "type": "integer",
                "format": "int64",
                "x-go-name": "Points"
              },
              "question": {
                "type": "string",
                "x-go-name": "Question"
              },
              "question_id": {
                "type": "string",
                "x-go-name": "QuestionId"
              },
              "question_media": {
                "type": "string",
                "x-go-name": "QuestionsMedia"
              },
              "question_type": {
                "type": "string",
                "x-go-name": "QuestionType"
              },
              "question_type_id": {
                "type": "integer",
                "format": "int64",
                "x-go-name": "QuestionTypeID"
              },
              "raw_options": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "uint8"
                },
                "x-go-name": "RawOptions"
              },
              "resource": {
                "type": "string",
                "x-go-name": "Resource"
              }
            },
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ResponseInsertImage": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    },
    "ResponseListQuestionByQuizId": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/QuestionAnalytics"
                },
                "x-go-name": "Data"
              },
              "is_quiz_editable": {
                "type": "boolean",
                "x-go-name": "IsQuizEditable"
              },
              "permission": {
                "type": "string",
                "x-go-name": "Permission"
              },
              "quiz_played_count": {
                "type": "integer",
                "format": "int64",
                "x-go-name": "QuizPlayedCount"
              }
            },
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ResponseOkWithMessage": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    },
...
  }
}